% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/add_word_counts_table.R
\name{add_word_counts_table}
\alias{add_word_counts_table}
\title{Adds word counts to a data.frame of commits.}
\usage{
add_word_counts_table(commit.table, repo)
}
\arguments{
\item{commit.table}{Data.frame of commits, produced by extract_commits.}

\item{repo}{Character string giving the relative or absolute path to the git repository.}
}
\value{
Updated data.frame of commits with word counts as new columns.
}
\description{
Adds word counts to a data.frame of commits.
}
\examples{
a <- extract_commits(repo="~/repositories/project/")
a <- add_word_counts_table(a,repo="~/repositories/project/")
}

