% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_plots.R
\name{build_plots}
\alias{build_plots}
\title{Builds plots of net additions}
\usage{
build_plots(words.table, metric = "net")
}
\arguments{
\item{metric}{(optional) Character string indicating which word count measure should be plotted. Acceptable options are "net" (net additions and modifications) and "nd" (new and deleted)}

\item{Data.frame}{of word counts for each date (produced by collapse_date)}
}
\value{
A multiplot containing line graphs and density histograms of the chosen metric..
}
\description{
Builds plots of net additions
}
\examples{
a <- extract_commits(repo="~/repositories/project/")
a <- add_word_counts_table(a,repo="~/repositories/project/")
b <- collapse_date(a)
build_plots(b)
}

