% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/collapse_date.R
\name{collapse_date}
\alias{collapse_date}
\title{Collapses a data.frame of word counts by date}
\usage{
collapse_date(commit.table, range = c(-4000, 10000))
}
\arguments{
\item{commit.table}{Data.frame of commits, produced by extract_commits.}

\item{range}{Numeric vector of two elements that specifies the upper/lower bounds of word counts. Anything outside this range is considered an outlier and is dropped.}
}
\value{
A data.frame containing dates and the corresponding counts for new words, deleted words, and net additions.
}
\description{
Takes a table of word counts (produced by extract_commits and then modified by add_word_counts_table) and collapses the word counts by date.
}
\examples{
a <- extract_commits(repo="~/repositories/project/")
a <- add_word_counts_table(a,repo="~/repositories/project/")
b <- collapse_date(a)
}

