% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/count_words.R
\name{count_words}
\alias{count_words}
\title{Counts the number of new and deleted words between two git commits}
\usage{
count_words(commit_old, commit_new, repo)
}
\arguments{
\item{commit_old}{Character string giving the previous commit ID}

\item{commit_new}{Character string giving the current commit ID}

\item{repo}{Character string giving the relative or absolute path to the git repository.}
}
\value{
List containing two elements: number of new words, and number of deleted words.
}
\description{
Calls git diff from the shell in porcelain mode to calculate the number of new and deleted words between two commits.
}
\examples{
count_words(commit_old="35c964f569ce26ab2a9b35469252730c854e6cb2",commit_new="849b280bccd4e3b294bc4a8a67f9f80184009edf",repo="~/repositories/project/")
}

