% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_interday.R
\name{drop_interday}
\alias{drop_interday}
\title{Drops interday commits to prevent double counting.}
\usage{
drop_interday(commit.table)
}
\arguments{
\item{commit.table}{Data.frame of commits, produced by extract_commits.}
}
\value{
Data.frame of commits with interday entries removed.
}
\description{
Drops interday commits to prevent double counting.
}

