% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/extract_commits.R
\name{extract_commits}
\alias{extract_commits}
\title{Extracts a table of commits and dates/times from a git log}
\usage{
extract_commits(repo)
}
\arguments{
\item{repo}{Character string giving the relative or absolute path to the git repository.}
}
\value{
data.frame containing (a) a column of commits and (b) a column of dates/times in POSIXct format.
}
\description{
Extracts a table of commits and dates/times from a git log
}
\examples{
extract_commits(repo="~/repositories/project/")
}

