% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plot_words.R
\name{plot_words}
\alias{plot_words}
\title{Builds plots of net additions}
\usage{
plot_words(words.table, cols = 1)
}
\arguments{
\item{cols}{Numeric indicating the number of columns in the resulting plot}

\item{Data.frame}{of word counts for each date (produced by collapse_date)}
}
\value{
A multiplot containing (a) net additions as a function of time, and (b) a density histogram of net additions.
}
\description{
Builds plots of net additions
}
\examples{
a <- extract_commits(repo="~/repositories/project/")
a <- add_word_counts_table(a,repo="~/repositories/project/")
b <- collapse_date(a)
plot_words(b)
}

