% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/subset_commits_date.R
\name{subset_commits_date}
\alias{subset_commits_date}
\title{Subsets a commit.table to a specific range of dates.}
\usage{
subset_commits_date(commit.table, date)
}
\arguments{
\item{commit.table}{Data.frame of commits, produced by extract_commits.}

\item{date}{Character vector of dates. Can either be a single date, or a pair of dates representing a range.}
}
\value{
A data.frame of commits that is a subset of the original data.frame.
}
\description{
Subsets a commit.table to a specific range of dates.
}
\examples{
a <- extract_commits(repo="~/repositories/project/")
b <- subset_commits_date(a,date=c("2015-01-01","2015-02-20"))
c <- subset_commits_date(a,date="2015-01-01")
}

