% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weekly_stats.R
\name{weekly_stats}
\alias{weekly_stats}
\title{Function for calculating weekly statistics}
\usage{
weekly_stats(words.table)
}
\arguments{
\item{Data.frame}{of word counts for each date (produced by collapse_date)}
}
\value{
Matrix of weekly statistics
}
\description{
Function for calculating weekly statistics
}

