% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/word_stats.R
\name{word_stats}
\alias{word_stats}
\title{Function to compute summary statistics (called by the shell script)}
\usage{
word_stats(date = NULL, repo = getwd())
}
\arguments{
\item{date}{(optional) character vector of dates, can either be of length 1, 2, or left unspecified. If length(date)==1, then the summary statistics will include all commits from that day until the latest commit. If length(date)==2, then date will be treated as a range, and the summary statistics will include all commits within that range. If date is left unspecified, then the summary statistics will include all commits in the repository.}

\item{repo}{(optional) repo (optional) character string giving the relative or absolute path to the git repository. If left unspecified, the current working directory will be used as the repo.}
}
\value{
Matrix of summary statistics for word counts
}
\description{
Function to compute summary statistics (called by the shell script)
}
\examples{
word_stats()
word_stats(date="2015-01-01")
word_stats(date="2015-01-01",repo="~/repositories/project/")
word_stats(date=c("2015-01-01","2015-02-20"))
word_stats(date=c("2015-01-01","2015-02-20"),repo="~/repositories/project/")
}

