% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/build_plots.R
\name{build_plots}
\alias{build_plots}
\title{Function to builds plots (called by the shell script)}
\usage{
build_plots(date = NULL, repo = getwd())
}
\arguments{
\item{date}{(optional) character vector of dates, can either be of length 1, 2, or left unspecified. If length(date)==1, then the plots will include all commits from that day until the latest commit. If length(date)==2, then date will be treated as a range, and the plot will include all commits within that range. If date is left unspecified, then the plot will include all commits in the repository.}

\item{repo}{(optional) character string giving the relative or absolute path to the git repository. If left unspecified, the current working directory will be used as the repo.}
}
\value{
Nothing returned - plots are saved to the directory specified in repo as 'wordplot.pdf'
}
\description{
Function to builds plots (called by the shell script)
}
\examples{
build_plots()
build_plots(date="2015-01-01")
build_plots(date="2015-01-01",repo="~/repositories/project/")
build_plots(date=c("2015-01-01","2015-02-20"))
build_plots(date=c("2015-01-01","2015-02-20"),repo="~/repositories/project/")
}

