% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_plots.R
\name{build_plots}
\alias{build_plots}
\title{Builds plots of net additions}
\usage{
build_plots(words.table, metric = "net", cols = 1)
}
\arguments{
\item{metric}{(optional) Character string indicating which word count measure should be plotted. Acceptable options are "net", "new", or "deleted".}

\item{cols}{(optional) Numeric indicating the number of columns in the resulting plot}

\item{Data.frame}{of word counts for each date (produced by collapse_date)}
}
\value{
A multiplot containing (a) net additions as a function of time, and (b) a density histogram of net additions.
}
\description{
Builds plots of net additions
}
\examples{
a <- extract_commits(repo="~/repositories/project/")
a <- add_word_counts_table(a,repo="~/repositories/project/")
b <- collapse_date(a)
build_plots(b)
}

