% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_words.R
\name{plot_words}
\alias{plot_words}
\title{Function to builds plots (called by the shell script)}
\usage{
plot_words(date = NULL, repo = getwd(), metric = "net", tofile = FALSE)
}
\arguments{
\item{date}{(optional) character vector of dates, can either be of length 1, 2, or left unspecified. If length(date)==1, then the plots will include all commits from that day until the latest commit. If length(date)==2, then date will be treated as a range, and the plot will include all commits within that range. If date is left unspecified, then the plot will include all commits in the repository.}

\item{repo}{(optional) character string giving the relative or absolute path to the git repository. If left unspecified, the current working directory will be used as the repo.}

\item{metric}{(optional) Character string indicating which word count measure should be plotted. Acceptable options are "net", "new", or "deleted".}

\item{tofile}{(optional) logical specifying whether the plot should be written to a pdf file (instead of displayed interactively)}
}
\description{
Function to builds plots (called by the shell script)
}
\examples{
plot_words()
plot_words(date="2015-01-01")
plot_words(date="2015-01-01",repo="~/repositories/project/")
plot_words(date=c("2015-01-01","2015-02-20"))
plot_words(date=c("2015-01-01","2015-02-20"),repo="~/repositories/project/")
}

